Agreement

+++ src/test/java/com/smalaca/rentalapplication/infrastructure/persistence/jpa/apartmentoffer/JpaApartmentOfferRepositoryIntegrationTest.java
@@ -2,6 +2,7 @@ package com.smalaca.rentalapplication.infrastructure.persistence.jpa.apartmentof
 
 import com.smalaca.rentalapplication.domain.apartmentoffer.ApartmentOffer;
 import com.smalaca.rentalapplication.domain.apartmentoffer.ApartmentOfferAssertion;
+import com.smalaca.rentalapplication.domain.apartmentoffer.ApartmentOfferRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
@@ -13,6 +14,7 @@ import java.time.LocalDate;
 import java.util.UUID;
 
 import static com.smalaca.rentalapplication.domain.apartmentoffer.ApartmentOffer.Builder.apartmentOffer;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Tag("DomainRepositoryIntegrationTest")
@@ -22,30 +24,71 @@ class JpaApartmentOfferRepositoryIntegrationTest {
     private static final LocalDate START = LocalDate.of(2040, 10, 11);
     private static final LocalDate END = LocalDate.of(2040, 10, 20);
 
-    @Autowired private JpaApartmentOfferRepository jpaApartmentOfferRepository;
+    @Autowired private ApartmentOfferRepository apartmentOfferRepository;
     @Autowired private SpringJpaApartmentOfferRepository springJpaApartmentOfferRepository;
 
     private UUID apartmentOfferId;
 
     @AfterEach
     void deleteApartmentOffer() {
-        springJpaApartmentOfferRepository.deleteById(apartmentOfferId);
+        if (apartmentOfferId != null) {
+            springJpaApartmentOfferRepository.deleteById(apartmentOfferId);
+        }
     }
 
     @Test
     void shouldSaveApartmentOffer() {
+        ApartmentOffer apartmentOffer = givenApartmentOffer();
+
+        apartmentOfferId = apartmentOfferRepository.save(apartmentOffer);
+
+        assertApartmentOffer(springJpaApartmentOfferRepository.findById(apartmentOfferId).get());
+    }
+
+    @Test
+    void shouldFindApartmentOfferByApartmentId() {
+        givenExistingApartmentOffer();
+
+        ApartmentOffer actual = apartmentOfferRepository.findByApartmentId(APARTMENT_ID);
+
+        assertApartmentOffer(actual);
+    }
+
+    @Test
+    void shouldRecognizeThereIsNoOfferForGivenApartmentId() {
+        boolean actual = apartmentOfferRepository.existByApartmentId(APARTMENT_ID);
+
+        assertThat(actual).isFalse();
+    }
+
+    @Test
+    void shouldRecognizeThereIsOfferForGivenApartmentId() {
+        givenExistingApartmentOffer();
+
+        boolean actual = apartmentOfferRepository.existByApartmentId(APARTMENT_ID);
+
+        assertThat(actual).isTrue();
+    }
+
+    private void givenExistingApartmentOffer() {
+        ApartmentOffer apartmentOffer = givenApartmentOffer();
+        apartmentOfferId = apartmentOfferRepository.save(apartmentOffer);
+    }
+
+    private void assertApartmentOffer(ApartmentOffer actual) {
+        ApartmentOfferAssertion.assertThat(actual)
+                .hasIdEqualTo(apartmentOfferId)
+                .hasApartmentIdEqualTo(APARTMENT_ID)
+                .hasPriceEqualTo(PRICE)
+                .hasAvailabilityEqualTo(START, END);
+    }
+
+    private ApartmentOffer givenApartmentOffer() {
         ApartmentOffer apartmentOffer = apartmentOffer()
                 .withApartmentId(APARTMENT_ID)
                 .withPrice(PRICE)
                 .withAvailability(START, END)
                 .build();
-
-        apartmentOfferId = jpaApartmentOfferRepository.save(apartmentOffer);
-
-        ApartmentOfferAssertion.assertThat(springJpaApartmentOfferRepository.findById(apartmentOfferId).get())
-                .hasIdEqualTo(apartmentOfferId)
-                .hasApartmentIdEqualTo(APARTMENT_ID)
-                .hasPriceEqualTo(PRICE)
-                .hasAvailabilityEqualTo(START, END);
+        return apartmentOffer;
     }
 }
